added dev get on project detail, deleted devController
