+++ controllers/developersController.js
@@ -7,5 +7,18 @@ module.exports = {
             .create(req.body)
             .then(dbModel => res.json(dbModel))
             .catch(err => res.status(422).json(err));
+    },
+    findAll: function(req, res) {
+        db.Developer
+          .find(req.query)
+          .sort({ date: -1 })
+          .then(dbModel => res.json(dbModel))
+          .catch(err => res.status(422).json(err));
+      },
+      findDev: function(req, res) {
+        db.Developer
+        .find(req.res.id)
+        .then(dbModel => res.json(dbModel))
+        .catch(err => res.status(422).json(err));
     }
 }
